# storage/api-docs/api-docs.yaml
...
openapi: "3.0.0"
info:
  title: "Upstart API"
  version: "1.0.0"
paths:
...
  /api/v1/categories:
    get:
      tags:
        - "Category"
      summary: "Get Categories"
      operationId: "Category"
      parameters:
        - name: "language_code"
          in: "query"
          schema:
            type: "string"
            example: en
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  properties:
                    id:
                      type: "integer"
                    parent_id:
                      type: "integer"
                    ordering:
                      type: "integer"
                    title:
                      type: "string"
                    icon:
                      type: "string"
                    child:
                      type: "array"
                      items:
                        properties:
                          id:
                            type: "integer"
                          parent_id:
                            type: "integer"
                          ordering:
                            type: "integer"
                          title:
                            type: "string"
                          icon:
                            type: "string"
                        type: "object"
                  type: "object"
        "401":
          description: "Unauthenticated"
...
  /api/v1/resources:
    get:
      tags:
        - "Resource"
      summary: "Get all resources"
      operationId: "Resources"
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            example: 10
        - name: page
          in: query
          schema:
            type: integer
            example: 2
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user_id:
                      type: integer
                    title:
                      type: string
                    path:
                      type: string
        "403":
          description: "Forbidden Error"
      security:
        - bearerAuth: [ ]
...
components:
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT